+++ E_Bill/src/main/java/com/service/BillService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import com.model.Bill;
 
 public interface BillService {
-	List<Bill> getNotPaidBills(int meterno);
+	List<Bill> getNotPaidBills(long meterno);
 
-	List<Bill> getAllByMeterno(int meterno);	
+	List<Bill> getAllByMeterno(long meterno);	
 	
 	List < Bill > getAllBills();
